\name{Alzheimer}
\docType{data}
\alias{Alzheimer}
\title{Genotype frequencies for 70 SNPs related to Alzheimer's disease}
\description{
   The dataframe contains the genotype frequencies MM, Mm and mm for the
   70 SNPs for both cases and controls. The data are taken from table
   7.11 in Laird & Lange.
}
\usage{data(Alzheimer)}
\format{A data frame containing 70 observations.}
\source{Laird, N. M. and Lange, C. Table 7.11, p. 124}
\references{
Laird, N. M. and Lange, C. (2011) The fundamentals of modern statistical genetics.
Springer.
}
\keyword{datasets}
     