\name{Glyoxalase}
\alias{Glyoxalase}
\docType{data}
\title{
Glyoxalase genotype data
}
\description{
Biallelic glyoxalase genotype data for 17 populations from India
}
\usage{data("Glyoxalase")}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{\code{AA}}{number of homozygote AA individuals}
    \item{\code{AB}}{number of heterozygote AB individuals}
    \item{\code{BB}}{number of homozygote BB individuals}
  }
}
\source{
Olson, J.M. (1993) Table 3.
}
\references{
Olson, J.M. (1993) Testing the Hardy-Weinberg law across strata. Annals of Human Genetics, 57,  pp. 291-295.
}
\examples{
data(Glyoxalase)
}
\keyword{datasets}
