\name{HapMapCHBChr1}
\docType{data}
\alias{HapMapCHBChr1}
\title{Genotype frequencies for 225 SNPs on chromosome 1 of the CHB population.}
\description{
   The dataframe contains the genotype frequencies in generic notation,
   AA, AB and BB the first 225 polymorphic SNPs without missing data on
   chromosome 1 of the Han Chinese in Beijing. The data are compiled
   from the HapMap project, phase 3.2, containing genotype information
   of 84 individuals.
}
\usage{data(HapMapCHBChr1)}
\format{A matrix containing 225 rows and 3 columns (AA, AB, BB).}
\source{http://hapmap.ncbi.nlm.nih.gov/}
\references{
The International HapMap Consortium (2007). A second generation human haplotype map of over 3.1 million SNPs
Nature 449, pp. 851--861.
}
\keyword{datasets}
     