\name{JPTsnps}
\alias{JPTsnps}
\docType{data}
\title{
Bi-allelic SNPs from a Japanese population stratified by gender
}
\description{
\code{JPTsnps} contains genotype counts for the two sexes of ten single
nucleotide polymorphisms of the Japanese (JPT) sample of the 1000
Genomes project, consisting of 56 males and 48 females.

The first three columns contain the male genotype counts, and the last
three columns contain the female genotype counts.
}
\usage{data("JPTsnps")}
\format{data frame}
\source{
  The \href{https://www.internationalgenome.org}{The 1000 genomes project}.
}
\references{
Puig, X., Ginebra, J. and Graffelman, J. (2019) Bayesian model selection
for the study of Hardy-Weinberg proportions and homogeneity of gender
allele frequencies.
}
\examples{
data(JPTsnps)
}
\keyword{datasets}
