\name{Markers}
\docType{data}
\alias{Markers}
\title{SNP data and intensities}
\description{
   The dataframe contains the genotypes of 3 SNPs and two allele
   intensities of 146 individuals. The first column is a GT polymorphism
   that has missing values for several individuals. The second and third
   column (iG and iG) are the allele intensities of this
   polymorphism. Column 4 and 5 are covariate SNPs (an AC and an AG
   polymorphism) that have no missing values.
}
\usage{data(Markers)}
\format{A data frame containing 146 rows and 5 columns}
\references{
Graffelman, J. (2013) Exploring bi-allelelic genetic markers: the
HardyWeinberg package. Journal of Statistical Software. 
}
\keyword{datasets}
     