\name{Mourant}
\docType{data}
\alias{Mourant}
\title{Genotype frequencies for blood group locus MN}
\description{
   The dataframe contains the genotype frequencies MM, MN and NN for the
   MN blood group locus for 216 populations. The data are taken from table 2.5 in Mourant et
   al., using only entries with a sample size of at least 500.
}
\usage{data(Mourant)}
\format{A data frame containing 216 observations.}
\source{Mourant et al, Table 2.5}
\references{
Mourant, A. E. and Kope\'c, A. C. and Domaniewska-Sobczak, K. (1976) The
Distribution of the Human Blood Groups and other Polymorphisms. Second edition.
Oxford University Press, London.
}
\keyword{datasets}
     