\name{NistSTRs}
\alias{NistSTRs}
\docType{data}
\title{
  NIST autosomal STR data 
}
\description{
 \code{NistSTRs} contains 29 autosomal microsatellites (STRs) of individuals of Caucasian ancestry. The two alleles of an individual are separated into two columns for each STR.
}
\usage{data("NistSTRs")}
\format{
  A data frame with 361 observations on the following 58 variables.
  \describe{
    \item{\samp{CSF1PO-1}}{First allele of CSF1PO}
    \item{\samp{CSF1PO-2}}{Second allele of CSF1PO}
    \item{\samp{D10S1248-1}}{First allele of D10S1248}
    \item{\samp{D10S1248-2}}{Second allele of D10S1248}
    \item{\samp{D12S391-1}}{First  allele of D12S391}
    \item{\samp{D12S391-2}}{Second allele of D12S391}
    \item{\samp{D13S317-1}}{First allele of D13S317}
    \item{\samp{D13S317-2}}{Second allele of D13S317}
    \item{\samp{D16S539-1}}{First allele of D16S539}
    \item{\samp{D16S539-2}}{Second allele of D16S539}
    \item{\samp{D18S51-1}}{First allele of D18S51}
    \item{\samp{D18S51-2}}{Second allele of D18S51}
    \item{\samp{D19S433-1}}{First allele of D19S433}
    \item{\samp{D19S433-2}}{Second allele of D19S433}
    \item{\samp{D1S1656-1}}{First allele of D1S1656}
    \item{\samp{D1S1656-2}}{Second allele of D1S1656}
    \item{\samp{D21S11-1}}{First allele of D21S11}
    \item{\samp{D21S11-2}}{Second allele of D21S11}
    \item{\samp{D22S1045-1}}{First allele of D22S1045}
    \item{\samp{D22S1045-2}}{Second allele of D22S1045}
    \item{\samp{D2S1338-1}}{First allele of D2S1338}
    \item{\samp{D2S1338-2}}{Second allele of D2S1338}
    \item{\samp{D2S441-1}}{First allele of D2S441}
    \item{\samp{D2S441-2}}{Second allele of D2S441}
    \item{\samp{D3S1358-1}}{First allele of D3S1358}
    \item{\samp{D3S1358-2}}{Second allele of D3S1358}
    \item{\samp{D5S818-1}}{First allele of D5S818}
    \item{\samp{D5S818-2}}{Second allele of D5S818}
    \item{\samp{D6S1043-1}}{First allele of D6S1043}
    \item{\samp{D6S1043-2}}{Second allele of D6S1043}
    \item{\samp{D7S820-1}}{First allele of D7S820}
    \item{\samp{D7S820-2}}{Second allele of D7S820}
    \item{\samp{D8S1179-1}}{First allele of D8S1179}
    \item{\samp{D8S1179-2}}{Second allele of D8S1179}
    \item{\samp{F13A01-1}}{First allele of F13A01}
    \item{\samp{F13A01-2}}{Second allele of F13A01}
    \item{\samp{F13B-1}}{First allele of F13B}
    \item{\samp{F13B-2}}{Second allele of F13B}
    \item{\samp{FESFPS-1}}{First allele of FESFPS}
    \item{\samp{FESFPS-2}}{Second allele of FESFPS}
    \item{\samp{FGA-1}}{First allele of FGA}
    \item{\samp{FGA-2}}{Second allele of FGA}
    \item{\samp{LPL-1}}{First allele of LPL}
    \item{\samp{LPL-2}}{Second allele of LPL}
    \item{\samp{Penta_C-1}}{First allele of Penta_C}
    \item{\samp{Penta_C-2}}{Second allele of Penta_C}
    \item{\samp{Penta_D-1}}{First allele of Penta_D}
    \item{\samp{Penta_D-2}}{Second allele of Penta_D}
    \item{\samp{Penta_E-1}}{First allele of Penta_E}
    \item{\samp{Penta_E-2}}{Second allele of Penta_E}
    \item{\samp{SE33-1}}{First allele of SE33}
    \item{\samp{SE33-2}}{Second allele of SE33}
    \item{\samp{TH01-1}}{First allele of TH01}
    \item{\samp{TH01-2}}{Second allele of TH01}
    \item{\samp{TPOX-1}}{First allele of TPOX}
    \item{\samp{TPOX-2}}{Second allele of TPOX}
    \item{\samp{vWA-1}}{First allele of vWA}
    \item{\samp{vWA-2}}{Second allele of vWA}
  }
}
\source{
    http://strbase.nist.gov
}
\references{
   Hill, C. et al. (2013) U.S. population data for 29 autosomal STR loci. Forensic science international: Genetics 497 7(3):e82-3.
}
\examples{
data(NistSTRs)
}
\keyword{datasets}
