\name{JPTsnps}
\alias{JPTsnps}
\docType{data}
\title{
Bi-allelic SNPs from a Japanese population 
}
\description{
\code{JPTsnps} contains genotype counts for the two sexes of ten single
nucleotide polymorphisms of the Japanese (JPT) sample of the 1000
Genomes project.
}
\usage{data("JPTsnps")}
\format{data frame}
\source{
  The \href{https://www.internationalgenome.org}{The 1000 genomes project}.
}
\references{
Puig, X., Ginebra, J. and Graffelman, J. (2019) Bayesian model selection
for the study of Hardy-Weinberg proportions and homogeneity of gender
allele frequencies.
}
\examples{
data(JPTsnps)
}
\keyword{datasets}
